QuestionController refactoring

+++ src/main/java/net/project/web/QuestionController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import net.project.domain.Question;
 import net.project.domain.QuestionRepository;
+import net.project.domain.Result;
 import net.project.domain.User;
 
 @Controller
@@ -46,25 +47,37 @@ public class QuestionController {
 	@GetMapping("/{id}")
 	public String show(@PathVariable Long id, Model model) {
 		//Question questeion = questionRepository.findOne(id);
-		model.addAttribute("question", questionRepository.findById(id).get());
+		model.addAttribute("question", questionRepository.findById(id).get()); // get
 		return "/qna/show";
 	}
 	
 	@GetMapping("/{id}/form")
 	public String updateForm(@PathVariable Long id, Model model, HttpSession session) {
-		try {
-			Question question = questionRepository.findById(id).get();
-			hasPermission(session, question);
-			model.addAttribute("question", question);
-			return "/qna/updateForm";
-		} catch(IllegalStateException e) {
-			model.addAttribute("errorMessage", e.getMessage());
-			return "/users/login";
+		Question question = questionRepository.findById(id).get();
+		Result result = valid(session, question);		
+		if (!result.isValid()) {
+			model.addAttribute("errorMessage", result.getErrorMessage());
+			return "/user/login";
 		}
+		
+		model.addAttribute("question", question);
+		return "/qna/updateForm";
 		
 	}
 	
-	private boolean hasPermission(HttpSession session, Question question) {
+	private Result valid(HttpSession session, Question question) {
+		if (!HttpSessionUtils.isLoginUser(session)) {
+			return Result.fail("로그인이 필요합니다.");
+		}
+		
+		User loginUser = HttpSessionUtils.getUserFromSession(session);
+		if (!question.isSameWriter(loginUser)) {
+			return Result.fail("자신이  쓴 글만 수정, 삭제가 가능합니다. ");
+		}
+		return Result.ok();
+	}
+	
+/*	private boolean hasPermission(HttpSession session, Question question) {
 		if (!HttpSessionUtils.isLoginUser(session)) {
 			throw new IllegalStateException("로그인이 필요합니다. ");
 		}
@@ -74,34 +87,34 @@ public class QuestionController {
 			throw new IllegalStateException("자신이 쓴 글만 수정, 삭제가 가능합니다. ");
 		}
 		return true;
-	}
+	} */
 	
 	@PutMapping("/{id}") 
 	public String update(@PathVariable Long id, String title, String contents, Model model, HttpSession session) {
-		try {
-			Question question = questionRepository.findById(id).get();
-			hasPermission(session, question);
-			question.update(title, contents);
-			questionRepository.save(question);
-			return String.format("redirect:/questions/%d",  id);
-		} catch(IllegalStateException e) {
-			model.addAttribute("errorMessage", e.getMessage());
-			return "/users/login";
+		Question question = questionRepository.findById(id).get();
+		Result result = valid(session, question);		
+		if (!result.isValid()) {
+			model.addAttribute("errorMessage", result.getErrorMessage());
+			return "/user/login";
 		}
+		
+		question.update(title, contents);
+		questionRepository.save(question);
+		return String.format("redirect:/questions/%d",  id);
 				
 	}
 	
 	@DeleteMapping("/{id}")
 	public String delete(@PathVariable Long id, Model model, HttpSession session) {
-		try {
-			Question question = questionRepository.findById(id).get();
-			hasPermission(session, question);
-			questionRepository.deleteById(id);
-			return "redirect:/";
-		} catch(IllegalStateException e) {
-			model.addAttribute("errorMessage", e.getMessage());
-			return "/users/login";
+		Question question = questionRepository.findById(id).get();
+		Result result = valid(session, question);		
+		if (!result.isValid()) {
+			model.addAttribute("errorMessage", result.getErrorMessage());
+			return "/user/login";
 		}
+		
+		questionRepository.deleteById(id);
+		return "redirect:/";
 		
 	}
